Link .la files first, this should avoid linking to old libclamav as in bb #931.
A linker warning on OpenBSD has shown the real cause of this:
/usr/bin/ld: warning: libclamunrar_iface.so.3.3, needed by
../libclamav/.libs/libclamav.so.4.1, may conflict with libclamunrar_iface.so.4.1

When -L/usr/local/lib is needed (because of either bz2, gmp, zlib)
it was the first on the linker command-line.
If multiple -L are on the linker line they don't override each-other,
all -L directories are searched.
So "-L/usr/local/lib -lgmp -L../libclamunrar/.libs/ -lclamunrar"
means to link to libclamunrar in /usr/local/lib first, which causes linking
to the old library.
If we pass the .la files first to the linker, we avoid linking to old libraries.



git-svn: trunk@3835