Must free engine on exit, otherwise valgrind incorrectly reports leaks.
This occurs because the JIT stores pointers mangled (stores an integer in low
bits of the pointer), thus valgrind can't see that pointers to that memory
region still exist, and reports it as definetely lost, instead of still
reachable.
The fix is to always free the engine, even if this is the parent after the fork.