Fix hash impl when replacing values.
We never remove keys from the hash, we only replace keys that expire due to LRU,
with the new keys. However the new key is not inserted in the same place as the
old one! (it depends where it hashes) So we need to mark the old key as deleted,
and not as empty. Otherwise future searches could miss keys that are in the
hash!