Fix bus error on Sparc.
Now that we have a struct with a 64-bit int member, we need to align
everything to 64-bits on sparc! The FRAG itself was aligned, however
the ->fake member was at 4-bytes distance (so half the time unaligned).
Fix this by adding a dummy 64-bit member in the union to force alignment to at
least 64-bit.

Also fix some compiler warnings from make distcheck.