Fix cli_unlink warnings (bb #1956).
fd == 0 is used to mark an invalid fd (not fd -1).
Test that tempfile is not NULL before calling cli_unlink
(that is the cause of the Bad address errors).
Also call cli_unlink when a virus is found. This isn't strictly
necessary since the file will get removed by the parent's rmdirs(),
but there is no point in leaving it behind.