darwin IS thread safe (bb #1534).
darwin* never defined CL_THREAD_SAFE, which was probably the cause for bug
I found this out when fixing the no pthreads case (bb #1897) and accidentally
broke it instead of fixing it. The Mac OS X buildbot failed.

This commit moves the CL_THREAD_SAFE definition to the same place as BUILD_CLAMD
(have_pthreads=yes) so no system with pthread is forgotten.
darwin has pthread in libc so no linker flags are needed (which is probably why
defining CL_THREAD_SAFE was forgotten)