Improve DB load time (bb #1860).
Avoid loading signatures that go into an offset-mode BM trie into the prefilter,
since this trie doesn't use the prefilter for BM anyway.
This improves load time, and filter performance.

Also set unroll to 4 in sha256.c, this seems to give best performance.

Overall this is a 16% improvement in DB load time.