Reduce memory waste due to alignment (~14% improvement).
Previously we were aligning all sizes to 8, and we had FRAG_OVERHEAD of 8
(because we wanted everything 8 byte aligned).
Now we only align as needed, and also only align a size once significantly
reducing memory waste.
FRAG_OVERHEAD is only 2, with up to 7 bytes of padding in worst case, but
usually much better.