Updated embedded copy of LLVM to version 2.8 (bb #2327).
This fixes code generation for CPUs without CMOV, so we can reenable JIT for
these.
Dropped ARM since it was not used (so now it is used on X86 and PPC), and
tests/. So python is no longer required to run 'make check'.
By default the embedded copy is used, if you want to use system's LLVM 2.8 lib
then use ./configure --enable-llvm --with-system-llvm=/path/to/llvm-config.