c4w: retry del_instance.
Caller can't wait for the proper scanner threads to finish for some reason.
Retry the del_instance once after 1 scanner thread finished (not necesarely the
one we got called for though).
If we still fail to delete Caller will retry after 2 seconds, and if we still
fail just go on and Uninit us (of course we'll fail there too, but its shutdown
anyway).

Also terminate monitor thread prior to freeing engine, because if engine can't
be freed we'd never even try to stop monitor thread.