fmap: reduce amount of mmap() calls in fmap_aging.
Scanning /dev/shm on an old kernel is serialized due to the additional mmap
contention that fmap_aging generates.
With this change it regains some parallelism (still not optimal), and it also
helps on newer kernels (less mmap calls is always good).
This patch makes just one mmap call for an entire range of pages,
if we try to unmap adjacent pages.