fix stack smashing on Win64.
size_t and ssize_t was used interchangibly, assuming they are the same size.
But size_t is 8 bytes and ssize_t 4 bytes long on Win64. They are only equal on
Win32.
We could redefine ssize_t to be same size as size_t, except we don't know what
that should be (long long is always 64-bit).
So just fix the types, fmap_need_off_once_len expect a size_t*, so declare
the variable as size_t instead of ssize_t.