clamav: add private fts() implementation
This file has been copied out of glibc and slightly modified.

The fts() implementation in current libc does not support LFS which
looks like a glibc (only) restriction. The code has been copied out of
glibc and the used functions (like fstat()) have been modified to use the
public exported libc function instead of the internal one. The libc
functions are are using the LFS version.
The fts() functions gained the _priv_ prefix so they do not clash with
glibc's functions.

This patch will be dropped once glibc's fts() implementations is LFS
safe which is glibc 2.23+, the bug is tracked in glibc's bugzilla as
11460 => https://sourceware.org/bugzilla/show_bug.cgi?id=11460

Patch-Name: clamav_add_private_fts_implementation.patch
Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>
Signed-off-by: Mickey Sola <msola@sourcefire.com>