fts: use AC_COMPILE_IFELSE instead AC_TRY_RUN
AC_TRY_RUN is obsolte and this is also bad for cross-compiling. The main
reason why I try get rid of it is because this code will segfault due to
the first NULL pointer.
In the end we what to figure out if the FTS implementation compiles with
LFS so there is no need to execute that. Just try to compile it.

Patch-Name: fts-no-use-AC_TRY_RUN.patch
Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>
Signed-off-by: Mickey Sola <msola@sourcefire.com>