Restore the correct OEP for aspack unpacked exes
For certain aspack versions, the code wasn't restoring the correct
entrypoint.  This had ramifications for unpacked file scanning,
since now an EP that points outside of a file makes the PE not
have .MDB sigs evaluated against it (a change from earlier versions,
but one which the MRT is OK with).