New tmp sub-dir per scan; JSON meta improvements
This commit improves the layout of the tmp file output and the JSON
metadata output when using the --leave-temps and --gen-json options.

For all scans, each scan target will get a unique tmp sub-directory. If
using --leave-temps, that subdir will include the basename of the
original file to make it easier to identify. Additionally, when using
--leave-temps option, all extracted objects will have their
subdirectories extracted in recursive subdirectories including filename
prefixes where available. When not using the --leave-temps option, the
layout of the tmp sub-directory will remain flat, so as to alleviate the
possibility of exceeding PATH_MAX.

The JSON metadata generated by the --gen-json option is now generated
for all file types, not just a select few. The format is also
pretty-printed for readability and now includes filenames and file paths
when available.

Also:

- Added missing ALLMATCH check when determining if bytecode hooks should
be run.

- Added cl_engine_get_str API to windows libclamav symbol export file.