Fix file access issue if in low privelege process
Removing problematic call to convert file descriptors to filepaths.

Added filename and tempfile names to scandesc calls in clamd.

Added a general scan option to treat the scan engine as unprivileged,
meaning that the scan engine will not have read access to the file.

Added check to drop a temp file for RAR's where the we don't have
read access to the filepath provided (i.e. unprivileged is set, or
access() check fails).