libclamav: Generate hash for each new fmap
Signature alerts on content extracted into a new fmap such as normalized
HTML resulted in checking FP signatures against the fmap's hash value
that was initialized to all zeroes, and never computed.

This patch seeks will enable FP signatures of normalized HTML files or
other content that is extracted to a new fmap to work.  This patch
doesn't resolve the issue that normal people will write FP signatures
targeting the original file, not the normalized file and thus won't
really see benefit from this bug-fix.

Additional work is needed to traverse the fmap recursion lists and
FP-check all parent fmaps when an alert occurs.  In addition, the HTML
normalization method of temporarily overriding the ctx->fmap instead of
increasing the recursion depth and doing ctx->fmap++/-- will need to be
corrected for fmap reverse recursion traversal to work.