libclamav: Fixes scanning of embedded fmaps
Specifically this fixes use of cli_map_scandesc().

The cli_map_scandesc() function used to override the current fmap
settings with a new size and offset, performing a scan of the embedded
content.  This broke the ability to iterate backwards through the fmap
recursion array when an alert occurs to check each map's hash for
whitelist matches.

In order to fix this issue, it needed to be possible to duplicate an
fmap header for the scan of the embedded file without duplicating the
actual map/data.  This wasn't feasible with the posix fmap handle
implementation where the fmap header, bitmap array, and memory map
were all contiguouus.  This commit makes it possible by extracting the
fmap header and bitmap array from the mmap region, using instead a
pointer for both the bitmap array and mmap/data.  The resulting posix
fmap handle implementation as a result ended up working more similarly
to existing the Windows implementation.

In addition to the above changes, this commit fixes:
- fmap recursion tracking for cli_scandesc()
- a recursion tracking issue in cli_scanembpe() error handling