Add EmbeddedObjects to JSON
The metadata projecties JSON structure isn't recording file types found
embedded within a file such as self-extracting (SFX) types and office
document types (DOCX, PPTX, etc).  This presents a problem...

At present there's no way to know if the current file has ended and a
few file is found tacked on to the end of the first file.  If there
were, we could simply check if the type found by the raw-scan exists
within the first file, or after.

If within the first, and the type is an archive then it's reasonable to
conclude we're either observing zip headers (for SFXZIP detections) or
other files that are not compressed.

If the type ISN'T found within the first file, then we definitely have
whole new file to parse and we should do so with cli_magic_scan()
rather than only using these embedded type scanners.

At present we can't ignore SFXZIP detections even if the original file
type is a ZIP because we may have found two ZIPs appended together to
evade detection (a legitimate trick).  As a consequence, we will
effectively parse every zip entry twice.  The same issue applies to
types found within non-compressed archives.

This commit adds an EmbeddedObjects list to the metadata JSON object so
that the existance of these types is noted.

Additionally, this commit removes the two-part int64 cli_jsonint64()
implementation as json_object_new_int64() should be available
everywhere and the macro to detect such support was never set.