Only create rfc2397 tmp directory when needed
HTML normalization creates a tmp directory for storing rfc2397 style
links.  The vast majority of html does not make use of rfc2397 and thus
an excess of empty tmp directories are generated.  This commit alters
behavior to only create the rfc2397 directory when required if it does
not already exist.