libclamav: Support for HFS+ compressed files
ClamAV doesn't handle compressed attribute for hfs+ file catalog
entries.  

This patch adds support for FLATE compressed files.

To accomplish this, we had to find and parse the root/header node
of the attributes file, if one exists. Then, parse the attribute map
to check if the compressed attribute exists. If compressed, parse the
compression header to determine how to decompress it. Support is
included for both inline compressed files as well as compressed
resource forks. 

Inflating inline compressed files is straightforward.

Inflating a compressed resource fork requires more work: 
- Find location and size of the resource.
- Parse the resource block table.
- Inflate and write each block to a temporary file to be scanned.

Additional changes needed for this work:
- Make hfsplus_fetch_node work for both catalog and attributes.
- Figure out node size.
- Handle nodes that span several blocks.
- If the attributes are missing, or invalid, extraction continues.
  This behavior is to support malformed files which would also
  extract on macOS and perhaps other systems.

This patch also:
- Adds filename extraction for the hfs+ parser.
- Skips embedded file type detection for GPT image file types. This
  prevents double extraction of embedded files, or misclassfication
  of GPT images as MHTML, for example. This resolves bb12335.