bb12282: Fix clamd unix socket dir creation perms
If the clamd.conf enables the LocalSocket option and sets the unix
socket file in a directory that does not exist, clamd creates the
missing directory but with invalid 000 permissions bits, causing socket
creation to fail.

This patch sets the umask temporarily to allow creation of the
directory w/ dwrxwr-wr- (766) permissions.