Reduce likelihood of tmp file name collisions
In regression testing we observed occasional errors creating files or
directories. It appears as though these are caused by tmp files that use
identical prefixes. Such prefixed tmp file names currently have a
5-character hash suffix. Though these temporary files or directories are
deleted when no longer needed, there is still the possibility that there
could be a hash collision which causes the scan to error out.

This patch doubles the size of the short-hash to 10-characters to reduce
the chance of a collision.